% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-tree.R
\name{aggregation_tree}
\alias{aggregation_tree}
\title{Aggregation Trees}
\usage{
aggregation_tree(cates, X, maxdepth, cp = 0)
}
\arguments{
\item{cates}{CATEs vector.}

\item{X}{Covariate matrix (no intercept).}

\item{maxdepth}{The maximum depth of the tree (i.e, the maximum number of nodes connecting the root to the leaves).}

\item{cp}{Any split that does not decrease the overall lack of fit by a factor of cp is not attempted. Default is zero, meaning that trees are fully grown up to \code{maxdepth}.}
}
\value{
The tree, as a \code{\link[rpart]{rpart}} object.
}
\description{
Grows an aggregation tree.
}
\seealso{
\code{\link{subtree}}, \code{\link{plot_tree}}
}
\author{
Riccardo Di Francesco
}
