% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-trees.R
\name{analyze_aggtree}
\alias{analyze_aggtree}
\title{Analyzing Aggregation Trees}
\usage{
analyze_aggtree(
  object,
  n_groups,
  method = "aipw",
  scores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{An \code{aggTrees} object.}

\item{n_groups}{Number of desired groups.}

\item{method}{Either \code{"raw"} or \code{"aipw"}, controls how GATEs are estimated.}

\item{scores}{Optional, vector of scores to be used in estimating GATEs. Useful to save computational time if scores have already been estimated. Ignored if \code{method == "raw"}.}

\item{verbose}{Logical, whether to print in console.}
}
\value{
The fitted model, as an \code{\link[estimatr]{lm_robust}} object, and the scores (if \code{method == "raw"}, this is
\code{NULL}). Additionally, it prints LATEX code in the console if \code{verbose == TRUE} (the default).
}
\description{
For a user-specified number of groups, select the optimal grouping and obtain point estimation and inference about
the GATEs. Additionally, compute the average characteristics of the units in each group.
}
\details{
Aggregation trees are a three-step procedure. First, CATEs are estimated using any estimator. Second, a tree is grown
to approximate the CATEs. Third, the tree is pruned to derive a nested sequence of optimal groupings, one for each
granularity level. The user can then choose her preferred level of granularity to obtain point estimation and inference
about the GATEs. \cr

\code{analyze_aggtree} takes as input an \code{aggTrees} object constructed by \code{\link{build_aggtree}}. Then, for the
desired granularity level, chosen via the \code{n_groups} argument, it provides point estimation and inference about the
GATEs, together with the average characteristics of the units in each group. Notice that inference is valid only if the
honest sample is non-empty, that is, the user allocated some units to the honest sample when running
\code{\link{build_aggtree}}.\cr

The \code{method} argument controls how GATEs are estimated. If \code{"method" == "raw"}, we estimate via OLS the following
linear model:

\deqn{Y_i = \sum_{l = 1}^{|T|} L_{i, l} \gamma_l + \sum_{l = 1}^{|T|} L_{i, l} D_i \beta_l + \epsilon_i}

with \code{L_{i, l}} a dummy variable equal to one if the i-th unit falls in the l-th group, and \code{|T|} the
number of groups. If the treatment is randomly assigned, one can show that the estimated betas identify the GATEs of
each group. Thus, we can interpret the OLS results as usual. However, in observational studies these estimates are biased
due to selection into treatment. To get unbiased estimates, we can set \code{"method"} to \code{"aipw"} to construct
doubly-robust scores \code{y_i^*} and use them as a pseudo-outcome in the following regression:

\deqn{Y_i^* = \sum_{l = 1}^{|T|} L_{i, l} \beta_l + \epsilon_i}

This way, we get unbiased GATEs estimates, and we can again interpret OLS results as usual.  Honest regression forests
and 5-fold cross fitting are used to estimate the propensity score and the conditional mean function of the outcome.\cr

Regardless of the chosen \code{method}, GATEs are estimated using observations in the honest sample as defined by the
output of \code{\link{build_aggtree}}. If the honest sample is empty, the same data used to construct the tree are used to
estimate the models above. This invalidates inference and produces a warning message.
}
\references{
\itemize{
  \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
}
}
\seealso{
\code{\link{build_aggtree}}
}
\author{
Riccardo Di Francesco
}
