% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-estimation.R
\name{causal_ols_aggtree}
\alias{causal_ols_aggtree}
\title{Causal Effects Estimation of aggTrees Objects Via Linear Models}
\usage{
causal_ols_aggtree(object, y, X, D)
}
\arguments{
\item{object}{An \code{aggTrees} object.}

\item{y}{Outcome vector.}

\item{X}{Covariate matrix (no intercept).}

\item{D}{Treatment assignment vector.}
}
\value{
The fitted model, as a \code{\link[estimatr]{lm_robust}} object.
}
\description{
Uses the leaves of a tree stored in an \code{aggTrees} object to estimate a linear model via OLS. The estimated coefficients
correspond to the GATEs in each leaf. If the data used in the OLS estimation have not been used to grow the tree (a condition
called "honesty"), then standard errors for the tree's estimates are valid.
}
\details{
To get point estimates and standard errors of the GATEs, \code{causal_ols_aggtree} fits via OLS the following model:

\deqn{y_i = \sum_{l = 1}^{|T|} L_{i, l} \gamma_l + \sum_{l = 1}^{|T|} L_{i, l} D_i \beta_l + \epsilon_i}

with \code{L_{i, l}} a dummy variable equal to one if the i-th unit falls in the l-th leaf of the tree, and \code{|T|} the
number of leaves. It is immediate to notice that the estimated betas corresponds to the GATEs of the l-th group:

\deqn{E[Y | D = 1, L_l = 1] - E[Y | D = 0, L_l = \] = \gamma_l + \beta_l - \gamma_l = \beta_l}

Thus, standard errors on the estimated betas are standard errors on the estimated GATEs.\cr

Notice that "honesty" is a necessary requirement to get valid standard errors. Thus, observations in \code{y}, \code{X} and
\code{D} must not have been used to grow the tree or estimate the cates.\cr

Standard errors are estimated via the Eicker-Huber-White estimator.
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{estimate_aggtree}}
}
\author{
Riccardo Di Francesco
}
