% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-tree.R
\name{cross_validated_tree}
\alias{cross_validated_tree}
\title{Cross-Validated Tree}
\usage{
cross_validated_tree(tree)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}
}
\value{
The cross-validated tree, as a \code{\link[rpart]{rpart}} object.
}
\description{
Uses the cross-validation criterion to select the "best" subtree.
}
\details{
\code{tree} should be deep enough to allow the criterion to explore more trees. \cr
}
\examples{
## Loading data (using random subsample provided by Matias Cattaneo).
dta <- haven::read_dta("http://www.stata-press.com/data/r13/cattaneo2.dta")

Y <- as.matrix(dta[, "bweight"])
D <- as.matrix(dta[, "mbsmoke"])
X_names <- c("bweight", "mbsmoke", "deadkids", "monthslb", "lbweight")
X <- as.matrix(dta[, !(colnames(dta) \%in\% X_names)])

## Splitting sample.
set.seed(1986)
n <- dim(dta)[1]
est_idx <- sample(1:n, n / 2, replace = FALSE)

X_est <- X[est_idx, ]
Y_est <- Y[est_idx]
D_est <- D[est_idx]

X_agg <- X[-est_idx, ]
Y_agg <- Y[-est_idx]
D_agg <- D[-est_idx]

## Estimating CATEs using only estimation sample.
cates_forest <- grf::causal_forest(X = X_est, Y = Y_est, W = D_est)

## Growing tree using only aggregation sample.
cates <- predict(cates_forest, newdata = X_agg)$predictions
tree <- aggregation_tree(cates, X_agg, maxdepth = 3, cp = 0.01)

## Plotting.
plot_aggregation_tree(tree)

## Cross-validation.
cv_tree <- cross_validated_tree(tree)
plot_aggregation_tree(cv_tree)

}
