% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-tree.R
\name{cv_subtree}
\alias{cv_subtree}
\title{Cross-Validated Tree}
\usage{
cv_subtree(tree)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}
}
\value{
The cross-validated tree, as a \code{\link[rpart]{rpart}} object.
}
\description{
Uses the cross-validation criterion to select the "best" subtree.
}
\details{
\code{tree} should be deep enough to allow the criterion to explore more trees. \cr
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{subtree}}, \code{\link{plot_tree}}
}
\author{
Riccardo Di Francesco
}
