% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honesty.R
\name{estimate_aggtree}
\alias{estimate_aggtree}
\title{Estimation of aggTrees Objects}
\usage{
estimate_aggtree(object, cates, X)
}
\arguments{
\item{object}{An \code{aggTrees} object.}

\item{cates}{Estimated CATEs.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
An \code{aggTrees} object.
}
\description{
Replace leaf predictions of a tree stored in an \code{aggTrees} object using external data.
}
\details{
\code{estimate_aggtree} replaces the leaf estimates of a tree as follows. First, it "pushes" all observations in
\code{X} down the tree and finds the leaves where they fall. Then, it computes the average value of \code{cates} in
each leaf. These values are the new estimates.\cr

If observations in \code{X} have not been used to construct the tree, then the new predictions are "honest".\cr

Due to coding limitations, \code{estimate_aggtree} replaces the estimates only in the leaves of a tree. The
internal nodes' estimates are not replaced, so the user should ignore them and focus on the leaves of the tree.\cr

To get standard errors on the tree's estimates, please use \code{\link{ols_aggtree}}.
}
\seealso{
\code{\link{ols_aggtree}}, \code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
