% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honesty.R
\name{honest_aggtree}
\alias{honest_aggtree}
\title{Honest aggTrees}
\usage{
honest_aggtree(object, cates, X)
}
\arguments{
\item{object}{An \code{aggTrees} object.}

\item{cates}{Estimated CATEs.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
An \code{aggTrees} object.
}
\description{
Replace leaf predictions of a tree stored in an \code{aggTrees} object using external data.
}
\details{
\code{honest_estimation} replaces the leaf estimates of a tree as follows. First, it "pushes" all observations in
\code{X} down the tree and finds the leaves where they fall. Then, it computes the average value of \code{cates} in
each leaf. These values are the new estimates, and are honest if observations in \code{X} have not been used to
construct the tree.\cr

Due to coding limitations, honest trees from \code{aggTrees} objects show honest estimates only in their leaves. The
internal nodes show non-honest estimates, so the user should ignore them and focus on the leaves of tree.\cr

To get standard errors on the tree's estimates, please use \code{\link{honest_ols_aggtree}}.
}
\seealso{
\code{\link{honest_ols}},\code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
