% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honesty.R
\name{honest_ols_aggtree}
\alias{honest_ols_aggtree}
\title{Honest OLS Estimation aggTrees}
\usage{
honest_ols_aggtree(object, cates, X)
}
\arguments{
\item{object}{An \code{aggTrees} object.}

\item{cates}{Estimated CATEs.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
The fitted model, as a \code{\link[stats]{lm}} object.
}
\description{
Uses the leaves of a tree stored in an \code{aggTrees} object to estimate a linear model via OLS. If the data used in
the OLS estimation have not been used to grow the tree, then standard errors for the tree's estimates are valid.
}
\details{
To get standard errors on the tree's estimates, \code{honest_ols} fits via OLS the following model:

\deqn{cates_i = \sum_{l = 1}^{|T|} L_l \beta_l + \epsilon_i}

with \code{L_l} the l-th leaf of the tree, and \code{|T|} the number of leaves. It is immediate to notice that the l-th
coefficient corresponds to the GATEs of the l-th group. Thus, standard errors on the estimated coefficients are standard
errors on the estimated GATEs.\cr

Notice that honesty is a necessary requirement to get valid standard errors. Thus, \code{X} must not have been used to
grow the tree or estimate the cates.\cr

Standard errors are estimated via the Eicker-Huber-White estimator.
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{honest_estimation}}
}
\author{
Riccardo Di Francesco
}
