% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honesty.R
\name{honest_rpart}
\alias{honest_rpart}
\title{Honest rpart}
\usage{
honest_rpart(tree, y_honest, X_honest)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}

\item{y_honest}{Outcome vector of the honest sample.}

\item{X_honest}{Covariate matrix (no intercept) of the honest sample.}
}
\value{
An honest tree, as a \code{\link[rpart]{rpart}} object.
}
\description{
Replace leaf predictions of a \code{\link[rpart]{rpart}} object using external data.
}
\details{
Due to coding limitations, the honest tree shows honest estimates only in its leaves. The internal nodes show
non-honest estimates, so the user should ignore them and focus on the leaves of tree.\cr

To get standard errors on the tree's estimates, please use \code{\link{honest_ols_rpart}}.
}
\seealso{
\code{\link{honest_ols_rpart}}
}
\author{
Riccardo Di Francesco
}
