% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance-measures.R
\name{log_ratio_sd}
\alias{log_ratio_sd}
\title{Log Ratio of Standard Deviations}
\usage{
log_ratio_sd(X, D)
}
\arguments{
\item{X}{Covariate matrix (no intercept).}

\item{D}{Treatment assignment vector.}
}
\value{
1xp data frame storing logarithm of the ratio of standard deviations of each covariate.
}
\description{
Computes a measure of the difference in the dispersion of the
covariate distributions across treatment arms.
}
\details{
Log ratio of standard deviations are computed as the logarithm of the ratio of the within-arm standard deviations.
}
\keyword{internal}
