% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart-functions.R
\name{node_membership}
\alias{node_membership}
\title{Node Membership}
\usage{
node_membership(tree, X, node)
}
\arguments{
\item{tree}{An \code{\link[rpart]{rpart}} object.}

\item{X}{Covariate matrix (no intercept).}

\item{node}{Number of node.}
}
\value{
Logical vector denoting whether each observation in \code{X} falls into \code{node}.
}
\description{
Constructs a binary variable that encodes whether each observation falls into a particular node of an
\code{\link[rpart]{rpart}} object.
}
\examples{
## Generate data.
set.seed(1986)

n <- 3000
k <- 3

X <- matrix(rnorm(n * k), ncol = k)
colnames(X) <- paste0("x", seq_len(k))

Y <- exp(X[, 1]) + 2 * X[, 2] * X[, 2] > 0 + rnorm(n)

## Construct tree.
library(rpart)
tree <- rpart(Y ~ ., data = data.frame(Y, X))

## Extract number of leaves.
is_in_third_node <- node_membership(tree, X, 3)
head(is_in_third_node)

}
\seealso{
\code{\link{subtree}} \code{\link{leaf_membership}} \code{\link{get_leaves}}
}
\author{
Riccardo Di Francesco
}
