% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance-measures.R
\name{normalized_diff}
\alias{normalized_diff}
\title{Normalized Differences}
\usage{
normalized_diff(X, D)
}
\arguments{
\item{X}{Covariate matrix (no intercept).}

\item{D}{Treatment assignment vector.}
}
\value{
1xp data frame storing the normalized difference of each covariate.
}
\description{
Computes a measure of the difference between locations of the covariate distributions
across treatment arms.
}
\details{
Normalized differences are computed as the difference in the means of each covariate across treatment arms, normalized
by the sum of the within-arm variances.
}
\keyword{internal}
