% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honesty.R
\name{ols_rpart}
\alias{ols_rpart}
\title{OLS Estimation of rpart Objects}
\usage{
ols_rpart(tree, y, X)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}

\item{y}{Outcome vector.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
The fitted model, as a \code{\link[estimatr]{lm_robust}} object.
}
\description{
Uses the leaves of a tree stored in an \code{\link[rpart]{rpart}} object to estimate a linear model via OLS.
If the data used in the OLS estimation have not been used to grow the tree, then standard errors for the tree's
estimates are valid.
}
\details{
To get standard errors on the tree's estimates, \code{ols_rpart} fits via OLS the following model:

\deqn{y_i = \sum_{l = 1}^{|T|} L_l \beta_l + \epsilon_i}

with \code{L_l} the l-th leaf of the tree, and \code{|T|} the number of leaves. It is immediate to notice that the l-th
coefficient corresponds to the estimated conditional expectation of \code{y_i} in the l-th leaf. Thus, standard errors
on the estimated coefficients are standard errors on these estimates.\cr

Notice that honesty is a necessary requirement to get valid standard errors. Thus, \code{y_honest} and {X_honest} must
not have been used to grow the tree.\cr

Standard errors are estimated via the Eicker-Huber-White estimator.
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{estimate_rpart}}
}
\author{
Riccardo Di Francesco
}
