% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-importance.R
\name{plot_importance}
\alias{plot_importance}
\title{Variable Importance Plot}
\usage{
plot_importance(importance, k = 0, title = "")
}
\arguments{
\item{importance}{The output of \code{\link{var_importance}}.}

\item{k}{Display only the k most important covariates. All covariates are displayed by default.}

\item{title}{String. The title of the plot.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object.
}
\description{
Plots the relative variable importance.
}
\seealso{
\code{\link{var_importance}}
}
\author{
Riccardo Di Francesco
}
