% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-visualization.R
\name{plot_tree}
\alias{plot_tree}
\title{Tree-Plot}
\usage{
plot_tree(tree, leaves = get_leaves(tree), palette = "auto", ...)
}
\arguments{
\item{tree}{A  \code{\link[rpart]{rpart}} object.}

\item{leaves}{Number of leaves of the desired tree. This can be used to plot subtrees of \code{tree}.}

\item{palette}{Palette to color the nodes. This must be a vector of colors.}

\item{...}{Further arguments for \code{\link[rpart.plot]{prp}}. See the details section for suggested parameters.}
}
\value{
None. It plots a tree.
}
\description{
Plots a \code{\link[rpart]{rpart}} object.
}
\details{
Please refer to \code{\link[colorspace]{choose_palette}} for high-quality palettes.\cr

Among the most useful parameters to be passed in \code{...} there is:
  \item{\code{main}}{String for the title of the plot.}
  \item{\code{cex.main}}{Size of the title.}
  \item{\code{split.cex}}{Size of splitting labels.}
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{subtree}}, \code{\link{plot_tree_sequence}}, \code{\link{recursive_partitioning_plot}}
}
\author{
Riccardo Di Francesco
}
