% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-visualization.R
\name{plot_tree_sequence}
\alias{plot_tree_sequence}
\title{Tree-Plot sequence}
\usage{
plot_tree_sequence(tree, palette = "auto", ...)
}
\arguments{
\item{tree}{A  \code{\link[rpart]{rpart}} object.}

\item{palette}{Palette to color the nodes. this must be a vector of colors.}

\item{...}{Further arguments for \code{\link[rpart.plot]{prp}}. See the details section for suggested parameters.}
}
\value{
None. It plots a sequence of subtree.
}
\description{
Plots a sequence of \code{\link[rpart]{rpart}} subtrees.
}
\details{
Please refer to \code{\link[colorspace]{choose_palette}} for high-quality palettes.\cr

Among the most useful parameters to be passed in \code{...} there is:
\describe{
  \item{\code{main}}{String for the title of the plot.}
  \item{\code{cex.main}}{Size of the title.}
  \item{\code{split.cex}}{Size of splitting labels.}
}
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{subtree}}, \code{\link{plot_tree}}, \code{\link{recursive_partitioning_plot}}
}
\author{
Riccardo Di Francesco
}
