% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance-measures.R
\name{print.phased}
\alias{print.phased}
\title{Print Method for \code{phased} objects}
\usage{
\method{print}{phased}(x, latex = FALSE, ...)
}
\arguments{
\item{x}{A \code{phased} object.}

\item{latex}{Logical. If TRUE, prints latex code for the table.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
None. It prints a table with summary measures. Latex output can be requested by setting
\code{latex = TRUE}.
}
\description{
Prints a nice table displaying balance measures.
}
\details{
Latex code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}.
}
