% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_latex}
\alias{rename_latex}
\title{Renaming Variables for LATEX Usage}
\usage{
rename_latex(names)
}
\arguments{
\item{names}{string vector.}
}
\value{
The renamed string vector. Strings where "_" is not found are not modified by \code{rename_latex}.
}
\description{
Renames variables where the character "_" is used, which causes clashes in LATEX.
}
\keyword{internal}
