% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_split}
\alias{sample_split}
\title{Sample Splitting}
\usage{
sample_split(n, training_frac = 0.5)
}
\arguments{
\item{n}{Size of the sample to be split.}

\item{training_frac}{Fraction of units for the training sample.}
}
\value{
A list storing the indexes for the two different subsamples.
}
\description{
Splits the sample into training and honest subsamples.
}
\author{
Riccardo Di Francesco
}
