% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{subtree_rpart}
\alias{subtree_rpart}
\title{rpart Subtree}
\usage{
subtree_rpart(tree, leaves = NULL, cv = FALSE)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}

\item{leaves}{Number of leaves of the desired subtree. Default is number of leaves of full tree.}

\item{cv}{If \code{TRUE}, \code{leaves} is ignored and a cross-validation criterion is used to select a partition.}
}
\value{
The subtree, as a \code{\link[rpart]{rpart}} object.
}
\description{
Extracts a subtree with a user-specified number of leaves from an \code{\link[rpart]{rpart}} object.
}
\seealso{
\code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
