% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary.aggTrees}
\alias{summary.aggTrees}
\title{Summary Method for aggTrees Objects}
\usage{
\method{summary}{aggTrees}(object, ...)
}
\arguments{
\item{object}{\code{aggTrees} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes an \code{aggTrees} object.
}
\seealso{
\code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
