% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-importance.R
\name{var_importance}
\alias{var_importance}
\title{Variable Importance for Estimated CATEs}
\usage{
var_importance(cates, X)
}
\arguments{
\item{cates}{CATEs vector.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
A data frame storing the relative importance of each covariate.
}
\description{
Grows a regression forest using the estimated CATEs as the dependent variable,
and provides the covariates' relative importance.
}
\details{
\code{var_importance} fits a regression forest using the covariates in \code{X} and \code{cates} as the
dependent variable. This allows to gain some knowledge on how covariates are related to treatment effect
heterogeneity.\cr

Variable importance is computed as a weighted sum of how many times a particular covariate was split on
at each depth in the forest. For more details, please see \code{\link[grf]{variable_importance}}.\cr

Notice that variable importance is generally a rough diagnostic of how treatment effects relate to the
covariates. If two covariates are highly correlated, trees generally split on only one of those covariates, thus
assigning low importance to the other. Therefore, one should not conclude that covariates with low importance are
not related to heterogeneity. A more systematic way to assess how treatment effects relate to the covariates consists
of investigating how the average characteristics of the units vary across subpopulations that differ in the magnitude
of their treatment effects (see \code{\link{avg_characteristics_aggtree}}).
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{avg_characteristics_aggtree}}
}
\author{
Riccardo Di Francesco
}
