% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{aggregation_tree}
\alias{aggregation_tree}
\title{Aggregation Trees}
\usage{
aggregation_tree(cates, X, maxdepth = 3, ...)
}
\arguments{
\item{cates}{Estimated CATEs.}

\item{X}{Covariate matrix (no intercept).}

\item{maxdepth}{The maximum depth of the tree.}

\item{...}{Further arguments from \code{\link[rpart]{rpart.control}}.}
}
\value{
An \code{aggTrees} object.
}
\description{
Grows an aggregation tree for discovering heterogeneous subpopulations that differ in the magnitude
of their treatment effects.
}
\details{
Aggregation trees are a three-step procedure. First, CATEs are estimated using any estimator. Second, a tree is grown
to approximate the CATEs. Third, the tree is pruned to derive a nested sequence of optimal partitions.

\code{aggregation_tree} uses \code{X} to grow a tree approximating the CATEs provided by the user. \code{X} must
consists of a covariate matrix with variables that coincide or are a subset of those used to construct \code{cates}.

The tree is grown up to some stopping criteria that can be specified by the user. Please refer to
the \code{\link[rpart]{rpart.control}} documentation for this.\cr
}
\references{
\itemize{
  \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
}
}
\seealso{
\code{\link{estimate_aggtree}}, \code{\link{causal_ols_aggtree}}, \code{\link{plot.aggTrees}},
\code{\link{subtree_aggtree}}
}
\author{
Riccardo Di Francesco
}
