% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-tree.R
\name{aggregation_tree}
\alias{aggregation_tree}
\title{Aggregation Trees}
\usage{
aggregation_tree(cates, X, maxdepth, cp)
}
\arguments{
\item{cates}{CATEs vector.}

\item{X}{Covariate matrix (no intercept).}

\item{maxdepth}{The maximum depth (i.e, the maximum number of nodes connecting the root to the leaves) of the tree.}

\item{cp}{Minimum MSE increase to accept splits.}
}
\value{
A \code{\link[rpart]{rpart}} object.
}
\description{
Grows an aggregation tree.
}
\details{
For details about aggregation trees and their use, please see Di Francesco (2022). \cr
}
\examples{
## Loading data (using random subsample provided by Matias Cattaneo).
dta <- haven::read_dta("http://www.stata-press.com/data/r13/cattaneo2.dta")

Y <- as.matrix(dta[, "bweight"])
D <- as.matrix(dta[, "mbsmoke"])
X_names <- c("bweight", "mbsmoke", "deadkids", "monthslb", "lbweight")
X <- as.matrix(dta[, !(colnames(dta) \%in\% X_names)])

## Splitting sample.
set.seed(1986)
n <- dim(dta)[1]
est_idx <- sample(1:n, n / 2, replace = FALSE)

X_est <- X[est_idx, ]
Y_est <- Y[est_idx]
D_est <- D[est_idx]

X_agg <- X[-est_idx, ]
Y_agg <- Y[-est_idx]
D_agg <- D[-est_idx]

## Estimating CATEs using only estimation sample.
cates_forest <- grf::causal_forest(X = X_est, Y = Y_est, W = D_est)

## Growing tree using only aggregation sample.
cates <- predict(cates_forest, newdata = X_agg)$predictions
tree <- aggregation_tree(cates, X_agg, maxdepth = 3, cp = 0.01)

## Plotting.
plot_aggregation_tree(tree)

## It can be annoying selecting a palette each time.
palette <- colorspace::choose_palette()
plot_aggregation_tree(tree, palette)
plot_aggregation_tree(tree, palette)

## Displaying the whole sequence.
plot_aggregation_tree(tree, palette, sequence = TRUE)

}
