% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg-characteristics.R
\name{avg_characteristics_aggtree}
\alias{avg_characteristics_aggtree}
\title{Leaves Average Characteristics}
\usage{
avg_characteristics_aggtree(object, cates, X)
}
\arguments{
\item{object}{An \code{aggTrees} object.}

\item{cates}{Estimated CATEs.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
Prints LATEX code in the console.
}
\description{
Compute the average characteristics of units in each leaf of an \code{aggTrees} object.
}
\details{
The fact that a tree does not split on a particular covariate does not imply that such covariate is not related
to treatment effect heterogeneity. In fact, there are many ways in which groups with low or high treatment effect can be formed.
Looking at how the average characteristics of the units vary across the leaves of the tree is a more systematic way to assess 
how treatment effects relate to the covariates.\cr

To achieve this, \code{avg_characteristics_aggtree} regresses each covariate on a categorical variable denoting membership to 
a particular leaf. This way, we get the average characteristics of units in each leaf, together with a standard error.\cr

Standard errors are estimated via the Eicker-Huber-White estimator.
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{honest_ols}}
}
\author{
Riccardo Di Francesco
}
