% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg-characteristics.R
\name{avg_characteristics_rpart}
\alias{avg_characteristics_rpart}
\title{Leaves Average Characteristics}
\usage{
avg_characteristics_rpart(tree, X, gates_point = NULL, gates_sd = NULL)
}
\arguments{
\item{tree}{A \code{rpart} object.}

\item{X}{Covariate matrix (no intercept).}

\item{gates_point}{Standard errors for estimated GATEs, one for each leaf.}
}
\value{
Prints LATEX code in the console.
}
\description{
Compute the average characteristics of units in each leaf of an \code{\link[rpart]{rpart}} object.
}
\details{
\code{avg_characteristics_rpart} regresses each covariate on a set of dummies denoting leaf membership.
This way, we get the average characteristics of units in each leaf, together with a standard error. Standard errors are
estimated via the Eicker-Huber-White estimator.\cr

Compilation of the LATEX code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}.
}
\references{
\itemize{
  \item V Chernozhukov, M Demirer, E Duflo, I Fernandez-Val (2018). Generic Machine Learning Inference on Heterogeneous Treatment Effects in Randomized Experiments, with an application to immunization in India. arXiv preprint arXiv:1712.04802. \doi{10.48550/ARXIV.1712.04802}.
  \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
}
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{causal_ols_rpart}}, \code{\link{estimate_rpart}}
}
\author{
Riccardo Di Francesco
}
