% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg-characteristics.R
\name{avg_characteristics_rpart}
\alias{avg_characteristics_rpart}
\title{Leaves Average Characteristics}
\usage{
avg_characteristics_rpart(tree, y, X)
}
\arguments{
\item{tree}{A \code{rpart} object.}

\item{y}{Outcome vector.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
Prints LATEX code in the console.
}
\description{
Compute the average characteristics of units in each leaf of a \code{\link[rpart]{rpart}} object.
}
\details{
\code{avg_characteristics_rpart} regresses each covariate on a categorical variable denoting membership to
a particular leaf. This way, we get the average characteristics of units in each leaf, together with a standard error.\cr

Standard errors are estimated via the Eicker-Huber-White estimator.\cr

Compilation of the LATEX code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}.
}
\seealso{
\code{\link{ols_rpart}}, \code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
