% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance-measures.R
\name{balance_measures}
\alias{balance_measures}
\title{Balance Measures}
\usage{
balance_measures(X, D)
}
\arguments{
\item{X}{Covariate matrix (no intercept).}

\item{D}{Treatment assignment vector.}
}
\value{
Object of class \code{phased}.
}
\description{
Compute several balance measures useful to check whether the covariate distributions are balanced across
treatment arms.
}
\details{
For each covariate in \code{X}, \code{balance_measures} computes sample averages and standard deviations
for both treatment arms. Additionally, two balance measures are computed:
\describe{
  \item{\code{Norm. Diff.}}{Normalized differences, computed as the differences in the means of each covariate
  across treatment arms, normalized by the sum of the within-arm variances. They provide a measure of the
  discrepancy between locations of the covariate distributions across treatment arms.}
  \item{\code{Log S.D.}}{Log ratio of standard deviations are computed as the logarithm of the ratio of the
  within-arm standard deviations. They provide a measure of the
  discrepancy in the dispersion of the covariate distributions across treatment arms.}
  }
}
\examples{
## Loading data (using random subsample provided by Matias Cattaneo).
dta <- haven::read_dta("http://www.stata-press.com/data/r13/cattaneo2.dta")

Y <- as.matrix(dta[, "bweight"])
D <- as.matrix(dta[, "mbsmoke"])
X_names <- c("bweight", "mbsmoke", "deadkids", "monthslb", "lbweight")
X <- as.matrix(dta[, !(colnames(dta) \%in\% X_names)])

# Computing balance measures.
measures <- balance_measures(X, D)

# Printing nice table (set latex = TRUE for latex code).
print(measures)

}
