% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-trees.R
\name{build_aggtree}
\alias{build_aggtree}
\title{Building Aggregation Trees}
\usage{
build_aggtree(
  y,
  D,
  X,
  honest_frac = 0.5,
  method = "aipw",
  cates = NULL,
  is_honest = NULL,
  ...
)
}
\arguments{
\item{y}{Outcome vector.}

\item{D}{Treatment vector.}

\item{X}{Covariate matrix (no intercept).}

\item{honest_frac}{Fraction of observations to be allocated to honest sample.}

\item{method}{Either \code{"raw"} or \code{"aipw"}, controls how node predictions are computed.}

\item{cates}{Estimated CATEs. If not provided by the user, CATEs are estimated internally via a \code{\link[grf]{causal_forest}}.}

\item{is_honest}{Logical vector denoting which observations belong to the honest sample. Required only if the \code{cates} argument is used.}

\item{...}{Further arguments from \code{\link[rpart]{rpart.control}}.}
}
\value{
An \code{aggTrees} object.
}
\description{
Builds an aggregation tree for discovering heterogeneous subpopulations that differ in the magnitude
of their treatment effects.
}
\details{
Aggregation trees are a three-step procedure. First, CATEs are estimated using any estimator. Second, a tree is grown
to approximate the CATEs. Third, the tree is pruned to derive a nested sequence of optimal groupings, one for each
granularity level. The user can then choose her preferred level of granularity to obtain point estimation and inference
about the GATEs. \cr

\code{build_aggtree} constructs the sequence of groupings and estimate the GATEs in each node. To obtain inference,
please check \code{\link{analyze_aggtree}}. Notice that inference is valid only if the honest sample is non-empty.\cr

After we constructed the sequence of groupings, GATEs can be estimated in several ways. This is controlled by the
\code{method} argument. If \code{"method" == "raw"}, we compute the difference in mean outcomes between treated and
control observations in each node. This is an unbiased estimator in randomized experiment. If \code{"method" == "aipw"}, we
construct doubly-robust scores and average them in each node. This is unbiased also in observational studies. Honest
regression forests and 5-fold cross fitting are used to estimate the propensity score and the conditional mean function of
the outcome.\cr

Regardless of the chosen \code{method}, GATEs are estimated using observations in the honest sample. If the honest sample
is empty, the same data used to construct the tree are used to estimate GATEs. This is fine for prediction but invalidates
the inference obtained by \code{\link{analyze_aggtree}}.\cr

The user can provide a vector of estimated CATEs via the \code{cates} argument. If so, the user needs to specify a logical
vector to denote which observations belong to the honest sample. If honesty is not desired, \code{is_honest} must be a
vector of \code{FALSE}s. If no vector of CATEs is provided, these are estimated internally via a
\code{\link[grf]{causal_forest}}.\cr

The tree is grown up to some stopping criteria that can be specified by the user. Please refer to
the \code{\link[rpart]{rpart.control}} documentation for this.\cr
}
\references{
\itemize{
  \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
}
}
\seealso{
\code{\link{analyze_aggtree}} \code{\link{plot.aggTrees}}
}
\author{
Riccardo Di Francesco
}
