% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-trees.R
\name{build_aggtree}
\alias{build_aggtree}
\title{Building Aggregation Trees}
\usage{
build_aggtree(y, D, X, honest_frac = 0.5, cates = NULL, is_honest = NULL, ...)
}
\arguments{
\item{y}{Outcome vector.}

\item{D}{Treatment vector.}

\item{X}{Covariate matrix (no intercept).}

\item{honest_frac}{Fraction of observations to be allocated to honest sample.}

\item{cates}{Estimated CATEs. If not provided by the user, CATEs are estimated internally via a \code{\link[grf]{causal_forest}}.}

\item{is_honest}{Logical vector denoting which observations belong to honest sample. Required only if the \code{cates} argument is used.}

\item{...}{Further arguments from \code{\link[rpart]{rpart.control}}.}
}
\value{
An \code{aggTrees} object.
}
\description{
Builds an aggregation tree for discovering heterogeneous subpopulations that differ in the magnitude
of their treatment effects.
}
\details{
Aggregation trees are a three-step procedure. First, CATEs are estimated using any estimator. Second, a tree is grown
to approximate the CATEs. Third, the tree is pruned to derive a nested sequence of optimal groupings, one for each
granularity level. The user can then choose her preferred level of granularity to obtain point estimation and inference
about the GATEs. \cr

\code{build_aggtree} constructs the sequence of groupings. Although node estimates are provided and shown by the generic
S3 methods, these should be ignored by the user. To obtain point estimation and inference about the GATEs,
please check \code{\link{analyze_aggtree}}. Notice that inference is valid only if the honest sample is non-empty, that is,
\code{honest_frac} is non-zero.\cr

The user can provide a vector of estimated CATEs via the \code{cates} argument. If so, the user needs to specify a logical
vector to denote which observations belong to the honest sample. If honesty is not desired, \code{is_honest} must be a
vector of \code{FALSE}s. If no vector of CATEs is provided, these are estimated internally via a
\code{\link[grf]{causal_forest}}.\cr

The tree is grown up to some stopping criteria that can be specified by the user. Please refer to
the \code{\link[rpart]{rpart.control}} documentation for this.\cr
}
\references{
\itemize{
  \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
}
}
\seealso{
\code{\link{analyze_aggtree}}
}
\author{
Riccardo Di Francesco
}
