% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dr_scores}
\alias{dr_scores}
\title{Doubly-Robust Scores}
\usage{
dr_scores(y, D, X, k = 5)
}
\arguments{
\item{y}{Outcome vector.}

\item{D}{Treatment assignment vector.}

\item{X}{Covariate matrix (no intercept).}

\item{k}{Number of folds.}
}
\value{
A vector of scores.
}
\description{
Constructs doubly-robust scores via K-fold cross-fitting.
}
\details{
Honest regression forests are used to estimate the propensity score and the conditional mean function of the outcome.
}
\references{
\itemize{
  \item J Robins, A Rotnitzky (1995). Semiparametric efficiency in multivariate regression models with missing data. \doi{10.2307/2291135}.
  \item V Semenova, V Chernozhukov (2021). Debiased machine learning of conditional average treatment effects and other causal functions. \doi{10.1093/ectj/utaa027}.
}
}
\author{
Riccardo Di Francesco
}
