% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-estimation.R
\name{estimate_rpart}
\alias{estimate_rpart}
\title{Estimation of rpart Objects}
\usage{
estimate_rpart(tree, y, X)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}

\item{y}{Outcome vector.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
A tree with leaf predictions replaced, as a \code{\link[rpart]{rpart}} object.
}
\description{
Replace leaf predictions of a \code{\link[rpart]{rpart}} object using external data.
}
\details{
Due to coding limitations, \code{estimate_rpart} replaces the estimates only in the leaves of a tree. The
internal nodes' estimates are not replaced, so the user should ignore them and focus on the leaves of the tree.\cr

To get standard errors on the tree's estimates, please use \code{\link{causal_ols_rpart}}.
}
\seealso{
\code{\link{causal_ols_rpart}}
}
\author{
Riccardo Di Francesco
}
