% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-estimation.R
\name{estimate_rpart}
\alias{estimate_rpart}
\title{Estimation of rpart Objects}
\usage{
estimate_rpart(tree, X, y = NULL, D = NULL, cates = NULL, method = "raw")
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}

\item{X}{Covariate matrix (no intercept).}

\item{y}{Outcome vector.}

\item{D}{Treatment assignment vector.}

\item{cates}{Estimated CATEs.}

\item{method}{Either \code{"raw"} or \code{"cates"}, defines how leaf predictions are replaced.}
}
\value{
A tree with leaf predictions replaced, as a \code{\link[rpart]{rpart}} object.
}
\description{
Replace leaf predictions of a \code{\link[rpart]{rpart}} object using external data.
}
\details{
\code{estimate_rpart} replaces the leaf estimates of a tree as follows. First, it "pushes" all observations in
\code{X} down the tree and finds the leaves where they fall. Then, it replaces the predictions in each leaf according
to the user-specified \code{method}.

If \code{method = "raw"}, \code{estimate_rpart} replaces leaf predictions with the difference between the sample average of
the observed outcomes of treated units and the sample average of the observed outcomes of control units in each leaf, which is
an unbiased estimator of the GATEs if the assignment to treatment is randomized. This requires the user to specify \code{y} and
\code{D}.\cr

If \code{method = "cates"}, \code{estimate_rpart} replaces leaf predictions with the sample average of \code{cates} in each leaf.
This is a valid estimator of the GATEs in observational studies.\cr

\code{estimate_rpart} allows the user to implement "honest" estimation. If observations in \code{X} have not been used to construct
the tree, then the new predictions are honest in the sense of Athey and Imbens (2016). This allows the user to conduct valid inference
about the estimated GATEs with standard approaches, e.g., by constructing conventional confidence intervals. To get standard errors on
the tree's estimates, please use \code{\link{causal_ols_rpart}}.\cr

Due to coding limitations, \code{estimate_rpart} replaces the estimates only in the leaves of a tree. The
internal nodes' estimates are not replaced, so the user should ignore them and focus on the leaves of the tree.\cr
}
\references{
\itemize{
  \item S Athey, G Imbens (2016). Recursive partitioning for heterogeneous causal effects. Proceedings of the National Academy of Sciences. \doi{10.1073/pnas.1510489113}.
}
}
\seealso{
\code{\link{causal_ols_rpart}}
}
\author{
Riccardo Di Francesco
}
