% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-tree.R
\name{get_leaves}
\alias{get_leaves}
\title{Number of Leaves in Tree}
\usage{
get_leaves(tree)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}
}
\value{
The number of leaves.
}
\description{
Extracts the number of leaves of a given tree.
}
\details{
If the tree is just a root, then \code{get_leaves} returns 1.
}
