% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart-functions.R
\name{get_leaves}
\alias{get_leaves}
\title{Number of Leaves}
\usage{
get_leaves(tree)
}
\arguments{
\item{tree}{An \code{\link[rpart]{rpart}} object.}
}
\value{
The number of leaves.
}
\description{
Extracts the number of leaves of an \code{\link[rpart]{rpart}} object.
}
\seealso{
\code{\link{subtree}} \code{\link{node_membership}} \code{\link{leaf_membership}}
}
\author{
Riccardo Di Francesco
}
