% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{get_leaves}
\alias{get_leaves}
\title{Number of Leaves in Tree}
\usage{
get_leaves(tree)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}
}
\value{
The number of leaves.
}
\description{
Extracts the number of leaves of an \code{\link[rpart]{rpart}} object.
}
\seealso{
\code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
