% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart-functions.R
\name{leaf_membership}
\alias{leaf_membership}
\title{Leaf Membership}
\usage{
leaf_membership(tree, X)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
A factor whose levels denote in which leaf each unit falls.
}
\description{
Constructs a variable that encodes in which leaf of an \code{\link[rpart]{rpart}} object the units in a given data frame fall.
}
\seealso{
\code{\link{subtree}} \code{\link{node_membership}} \code{\link{get_leaves}}
}
\author{
Riccardo Di Francesco
}
