% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart-functions.R
\name{node_membership}
\alias{node_membership}
\title{Node Membership}
\usage{
node_membership(tree, X, node)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}

\item{X}{Covariate matrix (no intercept).}

\item{node}{Number of node.}
}
\value{
Logical vector denoting whether each observation in \code{X} falls into \code{node}.
}
\description{
Constructs a binary variable that encodes whether each observations falls into a particular node of an
\code{\link[rpart]{rpart}} object.
}
\seealso{
\code{\link{subtree}} \code{\link{leaf_membership}} \code{\link{get_leaves}}
}
\author{
Riccardo Di Francesco
}
