% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{plot.aggTrees}
\alias{plot.aggTrees}
\title{Plot Method for aggTrees Objects}
\usage{
\method{plot}{aggTrees}(
  x,
  leaves = get_leaves(x$tree),
  type = 2,
  palette = c("BuRd"),
  sequence = FALSE,
  ...
)
}
\arguments{
\item{x}{An \code{aggTrees} object.}

\item{leaves}{Number of leaves of the desired tree. This can be used to plot subtrees.}

\item{type}{Plotting style.}

\item{palette}{Palette to color the nodes.}

\item{sequence}{If \code{TRUE}, the whole sequence of optimal partitions is displayed in a short animation.}

\item{...}{Further arguments from \code{\link[rpart.plot]{prp}}.}
}
\description{
Plots an \code{aggTrees} object.
}
\details{
\code{palette} can be either a vector of colors, or a function that takes as an argument the number of nodes of a tree and
returns a vector of colors. Please refer to \code{\link[colorspace]{choose_palette}} for high-quality palettes.\cr
}
\references{
\itemize{
  \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
}
}
\seealso{
\code{\link{build_aggtree}}, \code{\link{analyze_aggtree}}
}
\author{
Riccardo Di Francesco
}
