% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-importance.R
\name{plot_importance}
\alias{plot_importance}
\title{Variable Importance Plot}
\usage{
plot_importance(importance, k = 0)
}
\arguments{
\item{importance}{The output of \code{\link{var_importance}}.}

\item{k}{Display only the k most important covariates. All covariates are displayed by default.}
}
\value{
None. It plots the relative importance of the covariates.
}
\description{
Plots the relative variable importance
}
\examples{
## Loading data (using random subsample provided by Matias Cattaneo).
dta <- haven::read_dta("http://www.stata-press.com/data/r13/cattaneo2.dta")

Y <- as.matrix(dta[, "bweight"])
D <- as.matrix(dta[, "mbsmoke"])
X_names <- c("bweight", "mbsmoke", "deadkids", "monthslb", "lbweight")
X <- as.matrix(dta[, !(colnames(dta) \%in\% X_names)])

## Splitting sample.
set.seed(1986)
n <- dim(dta)[1]
est_idx <- sample(1:n, n / 2, replace = FALSE)

X_est <- X[est_idx, ]
Y_est <- Y[est_idx]
D_est <- D[est_idx]

X_agg <- X[-est_idx, ]
Y_agg <- Y[-est_idx]
D_agg <- D[-est_idx]

## Estimating CATEs using only estimation sample.
cates_forest <- grf::causal_forest(X = X_est, Y = Y_est, W = D_est)

## Growing forest using only aggregation sample.
cates <- predict(cates_forest, newdata = X_agg)$predictions

## Extracting variable importance.
var_imp <- var_importance(cates, X_agg)
plot_importance(var_imp)

## Looking only at the ten most important covariates.
plot_importance(var_imp, k = 10)

}
