% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{print.aggTrees}
\alias{print.aggTrees}
\title{Print Method for aggTrees Objects}
\usage{
\method{print}{aggTrees}(x, ...)
}
\arguments{
\item{x}{\code{aggTrees} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints an \code{aggTrees} object.
}
\seealso{
\code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
