% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{print.phased}
\alias{print.phased}
\title{Print Method for \code{phased} objects}
\usage{
\method{print}{phased}(x, latex = FALSE, ...)
}
\arguments{
\item{x}{A \code{phased} object.}

\item{latex}{If TRUE, prints LATEX code for the table.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints a \code{phased} object.
}
\details{
Compilation of the LATEX code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}.
}
