% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-visualization.R
\name{recursive_partitioning_plot}
\alias{recursive_partitioning_plot}
\title{Recursive Partitioning Plots}
\usage{
recursive_partitioning_plot(
  tree,
  cates,
  X,
  low = "yellow",
  high = "red",
  size = 1
)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object. The tree must have been built using only two covariates.}

\item{cates}{CATEs vector.}

\item{X}{Covariate matrix (no intercept), the same used to construct \code{tree}.}

\item{low}{String. Color to represent more negative \code{cates}.}

\item{high}{String. Color to represent more positive \code{cates}.}

\item{size}{Size of points in the scatter plot.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object.
}
\description{
Plots the recursive partitioning of \code{\link[rpart]{rpart}} objects for a two-dimensional covariate space.
}
\details{
The plot is built as follows. First, a scatter plot of the two covariates is displayed. Each point is colored
according to the associated value of \code{cates}, so that it is immediate noticing where treatment effects
are stronger and lighter (colors can be specified by the user). Second, the axis-aligned splits of \code{tree}
are overimposed.
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{subtree}}, \code{\link{plot_tree}}
}
\author{
Riccardo Di Francesco
}
