% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_split}
\alias{sample_split}
\title{Sample Splitting}
\usage{
sample_split(
  n,
  estimation_frac = 0.5,
  aggregation_frac = 0.25,
  honest_frac = 0.25
)
}
\arguments{
\item{n}{Size of the sample to be split.}

\item{estimation_frac}{Fraction of units for the estimation sample.}

\item{aggregation_frac}{Fraction of units for the aggregation sample.}

\item{honest_frac}{Fraction of units for the honest sample.}
}
\value{
A list storing the indexes for the three different subsamples.
}
\description{
Splits the sample in three different parts: estimation sample, aggregation sample, and honest sample.
}
\details{
The estimation and the aggregation sample are mandatory to use \code{\link{aggregation_tree}}, meaning that both
\code{estimation_frac} and \code{aggregation_frac} must be non-zero.\cr

The honest sample is required only if the user wants to conduct valid inference, which generally comes at the price of
a larger mean squared error. If honesty is not desired, please set \code{honest_frac} equal to zero.
}
\seealso{
\code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
