% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_split}
\alias{sample_split}
\title{Sample Splitting}
\usage{
sample_split(n, estimation_frac = 0.5, honest_frac = 1 - estimation_frac)
}
\arguments{
\item{n}{Size of the sample to be split.}

\item{estimation_frac}{Fraction of units for the estimation sample.}

\item{honest_frac}{Fraction of units for the honest sample.}
}
\value{
A list storing the indexes for the two different subsamples.
}
\description{
Splits the sample in estimation sample and honest sample.
}
\details{
The honest sample is required only if the user wants to conduct valid inference, which generally comes at the price of
a larger mean squared error.
}
\seealso{
\code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
