% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-tree.R
\name{subtree}
\alias{subtree}
\title{Subtree}
\usage{
subtree(tree, leaves)
}
\arguments{
\item{tree}{A \code{\link[rpart]{rpart}} object.}

\item{leaves}{Number of leaves of the desired subtree.}
}
\value{
The subtree, as a \code{\link[rpart]{rpart}} object.
}
\description{
Extracts a subtree with a user-specified number of leaves.
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{subtree}}, \code{\link{plot_tree}}
}
\author{
Riccardo Di Francesco
}
