% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{subtree_aggtree}
\alias{subtree_aggtree}
\title{aggTrees Subtree}
\usage{
subtree_aggtree(x, leaves = NULL, cv = FALSE)
}
\arguments{
\item{x}{An \code{aggTrees} object.}

\item{leaves}{Number of leaves of the desired subtree. Default is number of leaves of full tree.}

\item{cv}{If \code{TRUE}, \code{leaves} is ignored and a cross-validation criterion is used to select a partition.}
}
\value{
An \code{aggTrees} object.
}
\description{
Extracts a subtree with a user-specified number of leaves from an \code{aggTrees} object.
}
\references{
\itemize{
  \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
}
}
\seealso{
\code{\link{aggregation_tree}}, \code{\link{subtree_rpart}}, \code{\link{plot.aggTrees}}
}
\author{
Riccardo Di Francesco
}
