% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-importance.R
\name{var_importance}
\alias{var_importance}
\title{Variable Importance for Estimated CATEs}
\usage{
var_importance(cates, X)
}
\arguments{
\item{cates}{CATEs vector.}

\item{X}{Covariate matrix (no intercept).}
}
\value{
A data frame storing the relative importance of each covariate.
}
\description{
Grows a regression forest using the estimated CATEs as the dependent variable,
and provides the covariates' relative importance.
}
\details{
\code{var_importance} fits a regression forest using the covariates in \code{X} and \code{cates} as the
dependent variable. This allows to gain some knowledge on how covariates are related to treatment effect
heterogeneity.\cr

Variable importance is computed as a weighted sum of how many times a particular covariate was split on
at each depth in the forest. For more details, please see \code{\link[grf]{variable_importance}}.
}
\seealso{
\code{\link{aggregation_tree}}
}
\author{
Riccardo Di Francesco
}
